pyproject.toml

[tool.poetry]
name = "protarrow"
version = "0.1.0"
description = ""
authors = ["0x26res <0x26res@gmail.com>"]
packages = [
    { include = "protarrow" }
]

[tool.poetry.dependencies]
googleapis-common-protos = "^1.53.0"
pyarrow = "^10.0.0"
protobuf = "^4.21.6"
python = "^3.8"


[tool.poetry.dev-dependencies]
black = "^22.10.0"
grpcio-tools = "^1.50.0"
isort = "^5.10.1"
pre-commit = "^2.20.0"
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
  target-version =  ['py39']

[tool.isort]
  profile = "black"
  src_paths = ["protarrow", "tests"]
  known_first_party = ["protarrow", "protarrow_protos" ]

[tool.bandit]
skips = ["B101", "B311"]
